requirements.txt
requests~=2.32.4
python-dotenv~=1.1.1
libvirt-python~=11.6.0